refactor: add deprecated prefix on useless code for next breaking release
