+++ src/DEPRECATED_actions/photos.ts
