+++ src/DEPRECATED_selectors/photos.ts
