+++ src/index.ts
@@ -1,7 +1,7 @@
 require('isomorphic-fetch')
 
-export * from './actions'
-export * from './reducers'
-export * from './selectors'
+export * from './DEPRECATED_actions'
+export * from './DEPRECATED_reducers'
+export * from './DEPRECATED_selectors'
 export * from './standalone'
 export * from './utils'
