+++ test/actions/life.test.ts
@@ -1,4 +1,4 @@
-import { putCredentials } from '../../src/actions/life'
+import { putCredentials } from '../../src/DEPRECATED_actions/life'
 
 describe('actions/life', () => {
   it(`should make a snapshot of putCredentials`, () => {
