+++ test/reducers/index.2.test.ts
@@ -1,4 +1,4 @@
-import { configureStore } from '../../src/reducers'
+import { configureStore } from '../../src/DEPRECATED_reducers'
 
 describe('reducers', () => {
   it('should take a snapshot of store instance', () => {
